requirements.txt
django-recurrence
django_compressor
django_extensions
django-pwa
django-webpush
Pillow
googlemaps
requests_oauthlib
opencv-python
better_profanity
fuzzywuzzy
django-richtextfield